# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: PawonWarga Backend API Service
  title: PawonWarga API
  version: "1.0"
paths:
...
  /dashboard/overview:
    get:
      description: 'Aggregated stats backing the main dashboard: headline, summary
        metrics, sentiment trend, per-platform volume, top keywords, topic distribution,
        and top sentiment drivers all combine posts AND comments (see repository.PostRepository''s
        Combined* methods) — unlike /mentions, which is post-only. recent_mentions
        is the one exception and stays post-only (a bare comment has no url/location/category
        of its own). topic_distribution/keyword_cloud/drivers are keyword-matched,
        not a stored classification — see service.ClassifyTopics/computeDrivers. trend.positive/neutral/negative
        are each bucket''s percentage share (0-100) and trend.*_count are that same
        bucket''s raw mention counts — the frontend''s Percentage/Total toggle switches
        between them (same shape as /sentiment/overview''s trend). The platform filter
        applies to everything except data_sources, which always reports every platform''s
        status regardless of the filter (it''s a connectivity widget, not a filtered
        breakdown). summary.*.change is always "-" — period-over-period comparison
        isn''t implemented. Response messages are localized via ?lang= or Accept-Language
        (id/en).'
      parameters:
      - description: Start date (YYYY-MM-DD, WIB) — omit for no lower bound
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD, WIB, inclusive) — omit for no upper bound
        in: query
        name: to
        type: string
      - description: X, Instagram, TikTok, News, or YouTube — omit for all platforms
        in: query
        name: platform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dashboard overview
      tags:
      - dashboard
...
securityDefinitions:
  ApiKeyAuth:
    description: Shared secret for the internal ingest routes (Python labeling worker)
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"